Improve model handling
Refactor it to add embeddings models too
