+++ src/agents/kyma/tools/search.py
@@ -2,38 +2,36 @@
 from langchain_core.pydantic_v1 import BaseModel
 from langchain_core.tools import tool
 
-from agents.kyma.tools.retriever import (
-    Retriever,
-    create_embedding_factory,
-    create_hana_connection,
-    openai_embedding_creator,
-)
-from utils.settings import DATABASE_PASSWORD, DATABASE_PORT, DATABASE_URL, DATABASE_USER, EMBEDDING_MODEL_DEPLOYMENT_ID
+from rag.retriever import HanaDBRetriever
+from utils.hana import create_hana_connection
+from utils.models.factory import ModelType, ModelFactory
+from utils.settings import DATABASE_PASSWORD, DATABASE_PORT, DATABASE_URL, DATABASE_USER, DOC_TABLE_NAME
+from langchain_core.embeddings import Embeddings
 
-create_embedding = create_embedding_factory(openai_embedding_creator)
-embeddings_model = create_embedding(EMBEDDING_MODEL_DEPLOYMENT_ID)
 # setup connection to Hana Cloud DB
-hana_conn = create_hana_connection(
-    DATABASE_URL, DATABASE_PORT, DATABASE_USER, DATABASE_PASSWORD
-)
-table_name = "kyma_os_docs"
 
 class SearchKymaDocArgs(BaseModel):
     """Arguments for the search_kyma_doc tool."""
 
     query: str
 
-
-@tool(infer_schema=False, args_schema=SearchKymaDocArgs)
-def search_kyma_doc_tool(query: str) -> str:
-    """Search through Kyma documentation for relevant information.
-    Provide a search query to find information about Kyma concepts, features, or components."""
+def create_search_kyma_doc_tool(embedding_model: Embeddings):
+  hana_conn = create_hana_connection(
+    DATABASE_URL, DATABASE_PORT, DATABASE_USER, DATABASE_PASSWORD
+  )
+  retriever = HanaDBRetriever(embedding_model, hana_conn, DOC_TABLE_NAME)
+  
+  @tool(infer_schema=False, args_schema=SearchKymaDocArgs)
+  def search_kyma_doc_tool(query: str) -> str:
+      """Search through Kyma documentation for relevant information.
+      Provide a search query to find information about Kyma concepts, features, or components."""
+      
+      docs = retriever.retrieve(query)
+      
+      if len(docs) == 0:
+          return "No relevant documentation found."
+      
+      docs_str = "\n".join([doc.page_content for doc in docs])
+      return docs_str
     
-    retriever = Retriever(embeddings_model, hana_conn, table_name)
-    docs = retriever.retrieve(query)
-    
-    if len(docs) == 0:
-        return "No relevant documentation found."
-    
-    docs_str = "\n".join([doc.page_content for doc in docs])
-    return docs_str
+  return search_kyma_doc_tool
