+++ src/initial_questions/inital_questions.py
@@ -10,7 +10,7 @@ from initial_questions.output_parser import QuestionOutputParser
 from initial_questions.prompts import INITIAL_QUESTIONS_PROMPT
 from services.k8s import IK8sClient
 from utils.logging import get_logger
-from utils.models import IModel
+from utils.models.factory import IModel
 from utils.utils import is_empty_str, is_non_empty_str
 
 logger = get_logger(__name__)
